Updated script

+++ scripts/package.py
@@ -408,39 +408,40 @@ def package_card_files(repo_root, files_root_dir, path_list, sdk_version):
 
         shutil.rmtree(os.path.join(repo_root, f'tmp/assets/{asset_type}/{mcu_card_name}'))
 
-    for each_query_path in os.listdir(os.path.join(os.getcwd(), 'resources/queries/cards')):
-        if each_query_path not in existing_cards:
-            print('Adding new card: %s' % each_query_path)
-            json_device = json.load(open(os.path.join(os.getcwd(), 'resources/queries/cards', each_query_path, 'Devices.json')))
-            card_path = json_device['uid'].rsplit('_', 1)[0].lower()
-            os.makedirs(os.path.join(repo_root, f'tmp/assets/{asset_type}/{each_query_path}'), exist_ok=True)
-            shutil.copyfile(
-                os.path.join(files_root_dir, card_path, 'mcu_card.h'),
-                os.path.join(repo_root, f'tmp/assets/{asset_type}/{each_query_path}/mcu_card.h')
-            )
-            package_name = json_device['uid'].lower()
-            create_custom_archive(
-                os.path.join(repo_root, f'tmp/assets/{asset_type}/{each_query_path}'),
-                os.path.join(repo_root, f'tmp/assets/{asset_type}/{package_name}.7z')
-            )
-            query_file = '\'{"package":"' + package_name + '"}\''
-            archive_list.update(
-                {
-                    json_device['name']:
+    if os.path.exists(os.path.join(os.getcwd(), 'resources/queries/cards')):
+        for each_query_path in os.listdir(os.path.join(os.getcwd(), 'resources/queries/cards')):
+            if each_query_path not in existing_cards:
+                print('Adding new card: %s' % each_query_path)
+                json_device = json.load(open(os.path.join(os.getcwd(), 'resources/queries/cards', each_query_path, 'Devices.json')))
+                card_path = json_device['uid'].rsplit('_', 1)[0].lower()
+                os.makedirs(os.path.join(repo_root, f'tmp/assets/{asset_type}/{each_query_path}'), exist_ok=True)
+                shutil.copyfile(
+                    os.path.join(files_root_dir, card_path, 'mcu_card.h'),
+                    os.path.join(repo_root, f'tmp/assets/{asset_type}/{each_query_path}/mcu_card.h')
+                )
+                package_name = json_device['uid'].lower()
+                create_custom_archive(
+                    os.path.join(repo_root, f'tmp/assets/{asset_type}/{each_query_path}'),
+                    os.path.join(repo_root, f'tmp/assets/{asset_type}/{package_name}.7z')
+                )
+                query_file = '\'{"package":"' + package_name + '"}\''
+                archive_list.update(
                     {
-                        "name": json_device['uid'].rsplit('_', 1)[0].lower(),
-                        "display_name": json_device['name'],
-                        "type": "card",
-                        "icon": f'https://raw.githubusercontent.com/MikroElektronika/mikrosdk_v2/master/resources/{json_device['icon']}',
-                        "package_name": package_name,
-                        "hash": hash_directory_contents(os.path.join(repo_root, f'tmp/assets/{asset_type}/{each_query_path}')),
-                        "category": "Card Package",
-                        "package_rel_path": f'tmp/assets/{asset_type}/{package_name}.7z',
-                        "install_location": f"%APPLICATION_DATA_DIR%/packages/sdk/mikroSDK_v2/src/bsp/board/include/mcu_cards/{card_path}/{json_device['def_file'].split('.')[0]}",
-                        "db_query": f'UPDATE Devices SET installer_package = {query_file} WHERE name = \"{json_device['name']}\"'
+                        json_device['name']:
+                        {
+                            "name": json_device['uid'].rsplit('_', 1)[0].lower(),
+                            "display_name": json_device['name'],
+                            "type": "card",
+                            "icon": f'https://raw.githubusercontent.com/MikroElektronika/mikrosdk_v2/master/resources/{json_device['icon']}',
+                            "package_name": package_name,
+                            "hash": hash_directory_contents(os.path.join(repo_root, f'tmp/assets/{asset_type}/{each_query_path}')),
+                            "category": "Card Package",
+                            "package_rel_path": f'tmp/assets/{asset_type}/{package_name}.7z',
+                            "install_location": f"%APPLICATION_DATA_DIR%/packages/sdk/mikroSDK_v2/src/bsp/board/include/mcu_cards/{card_path}/{json_device['def_file'].split('.')[0]}",
+                            "db_query": f'UPDATE Devices SET installer_package = {query_file} WHERE name = \"{json_device['name']}\"'
+                        }
                     }
-                }
-            )
+                )
 
     return archive_list
